+++ madminer/core.py
@@ -6,11 +6,11 @@ import logging
 from collections import OrderedDict
 import tempfile
 
-from madminer.morphing import PhysicsMorpher
+from madminer.utils.morphing import PhysicsMorpher
 from madminer.utils.interfaces.madminer_hdf5 import save_madminer_settings, load_madminer_settings
 from madminer.utils.interfaces.mg_cards import export_param_card, export_reweight_card, export_run_card
 from madminer.utils.interfaces.mg import generate_mg_process, setup_mg_with_scripts, run_mg, create_master_script
-from madminer.utils.various import create_missing_folders, format_benchmark, make_file_executable, copy_file
+from madminer.utils.various import create_missing_folders, format_benchmark, copy_file
 
 logger = logging.getLogger(__name__)
