+++ madminer/fisherinformation.py
@@ -7,7 +7,7 @@ import os
 
 from madminer.utils.interfaces.madminer_hdf5 import load_madminer_settings, madminer_event_loader
 from madminer.utils.analysis import get_theta_benchmark_matrix, get_dtheta_benchmark_matrix, mdot
-from madminer.morphing import PhysicsMorpher, NuisanceMorpher
+from madminer.utils.morphing import PhysicsMorpher, NuisanceMorpher
 from madminer.utils.various import format_benchmark, math_commands, weighted_quantile, sanitize_array
 from madminer.ml import Estimator, Ensemble
