+++ tests/test_imports.py
@@ -1,13 +1,2 @@
-from madminer.core import MadMiner
-from madminer.delphes import DelphesReader
-from madminer.lhe import LHEReader
-from madminer.ml import Ensemble, Estimator
-from madminer.morphing import PhysicsMorpher, NuisanceMorpher
-from madminer.plotting import plot_2d_morphing_basis, plot_distribution_of_information, plot_distributions
-from madminer.plotting import plot_fisher_information_contours_2d, plot_fisherinfo_barplot
-from madminer.plotting import plot_nd_morphing_basis_scatter, plot_2d_morphing_basis
-from madminer.sampling import SampleAugmenter
-
-
 def test_imports():
     assert True
